+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [0.3.1-SNAPSHOT]
+### Added
+- `navigate-fx` and `navigate` subscription
+
+### Changed
+- renamed `dom` namespace to `browser`
 
 ## [0.3.0]
 ### Added
 - focus-fx - in case you really need it
+
 ### Changed
 - going independent of re-frame